Moving the InferLegacyRestoreRuntime method to Test.Utilities, since it is needed only by tests.
