+++ test/EndToEnd/EndToEndTest.cs
@@ -5,7 +5,6 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
-using Microsoft.DotNet.Cli.Utils;
 using Microsoft.DotNet.InternalAbstractions;
 using Microsoft.DotNet.PlatformAbstractions;
 using Microsoft.DotNet.Tools.Test.Utilities;
@@ -197,7 +196,7 @@ namespace Microsoft.DotNet.Tests.EndToEnd
             TestProject = Path.Combine(TestDirectory, "project.json");
             OutputDirectory = Path.Combine(TestDirectory, s_outputdirName);
 
-            Rid = DotnetRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier();
+            Rid = DotnetLegacyRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier();
         }
 
         private static void SetupStaticTestProject()
