+++ test/EndToEnd/project.json
@@ -1,7 +1,6 @@
 {
   "version": "1.0.0-*",
   "buildOptions": {
-    "keyFile": "../../tools/Key.snk",
     "emitEntryPoint": true
   },
   "dependencies": {
