+++ test/Microsoft.DotNet.Cli.Utils.Tests/StreamForwarderTests.cs
@@ -13,7 +13,7 @@ namespace StreamForwarderTests
 {
     public class StreamForwarderTests : TestBase
     {
-        private static readonly string s_rid = DotnetRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier();
+        private static readonly string s_rid = DotnetLegacyRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier();
 
         public static void Main()
         {
