+++ test/dotnet-compile.UnitTests/project.json
@@ -33,6 +33,9 @@
     "Microsoft.DotNet.ProjectModel": {
       "target": "project"
     },
+    "Microsoft.DotNet.Tools.Tests.Utilities": {
+      "target": "project"
+    },
     "xunit": "2.2.0-beta3-build3330",
     "dotnet-test-xunit": "1.0.0-rc2-350904-49",
     "moq.netcore": "4.4.0-beta8",
