+++ test/dotnet-publish.Tests/PublishTests.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text.RegularExpressions;
-using Microsoft.DotNet.Cli.Utils;
 using Microsoft.DotNet.InternalAbstractions;
 using Microsoft.DotNet.TestFramework;
 using Microsoft.DotNet.Tools.Test.Utilities;
@@ -59,11 +58,11 @@ namespace Microsoft.DotNet.Tools.Publish.Tests
                 {
                     new object[] { "1", "", "", "", "" },
                     new object[] { "2", "netcoreapp1.0", "", "", "" },
-                    new object[] { "3", "", DotnetRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier(), "", "" },
+                    new object[] { "3", "", DotnetLegacyRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier(), "", "" },
                     new object[] { "4", "", "", "Release", "" },
                     new object[] { "5", "", "", "", "some/dir"},
                     new object[] { "6", "", "", "", "some/dir/with spaces" },
-                    new object[] { "7", "netcoreapp1.0", DotnetRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier(), "Debug", "some/dir" },
+                    new object[] { "7", "netcoreapp1.0", DotnetLegacyRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier(), "Debug", "some/dir" },
                 };
             }
         }
@@ -225,7 +224,7 @@ namespace Microsoft.DotNet.Tools.Publish.Tests
             publishCommand = new PublishCommand(
                 lesserTestProject,
                 "netcoreapp1.0",
-                DotnetRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier());
+                DotnetLegacyRuntimeIdentifiers.InferLegacyRestoreRuntimeIdentifier());
             publishCommand.Execute().Should().Pass();
 
             publishCommand.GetOutputDirectory().Should().HaveFile("TestLibraryLesser.dll");
