+++ test/dotnet-publish.Tests/project.json
@@ -1,7 +1,6 @@
 {
   "version": "1.0.0-*",
   "buildOptions": {
-    "keyFile": "../../tools/Key.snk",
     "copyToOutput": {
       "include": [
         "../../artifacts/*/stage2/sdk/*/.version"
@@ -21,9 +20,6 @@
       "target": "project"
     },
     "Microsoft.DotNet.Tools.Tests.Utilities": {
-      "target": "project"
-    },
-    "Microsoft.DotNet.Cli.Utils": {
       "target": "project"
     },
     "xunit": "2.2.0-beta3-build3330",
